feat: implement ContextMenu component and refactor song context menu usage
